Replace existing sidebar with new design
